requirements_base.txt
dj-database-url==0.2.2
gunicorn==18.0
static==1.0.2
otree-core==0.3.8
django-mathfilters==0.3.0
